ref(uptime): move helper functions to utility file
