+++ src/sentry/uptime/subscriptions/subscriptions.py
@@ -1,7 +1,6 @@
 import logging
 from collections.abc import Sequence
 
-from django.conf import settings
 from django.db import router, transaction
 from sentry_kafka_schemas.schema_types.uptime_results_v1 import (
     CHECKSTATUS_FAILURE,
@@ -40,7 +39,7 @@ from sentry.uptime.types import (
     GROUP_TYPE_UPTIME_DOMAIN_CHECK_FAILURE,
     UptimeMonitorMode,
 )
-from sentry.utils import redis
+from sentry.uptime.utils import build_fingerprint, build_last_update_key, get_cluster
 from sentry.utils.db import atomic_transaction
 from sentry.utils.not_set import NOT_SET, NotSet, default_if_not_set
 from sentry.utils.outcomes import Outcome
@@ -56,22 +55,6 @@ UPTIME_SUBSCRIPTION_TYPE = "uptime_monitor"
 MAX_AUTO_SUBSCRIPTIONS_PER_ORG = 1
 MAX_MANUAL_SUBSCRIPTIONS_PER_ORG = 100
 MAX_MONITORS_PER_DOMAIN = 100
-
-
-def build_last_update_key(detector: Detector) -> str:
-    return f"project-sub-last-update:detector:{detector.id}"
-
-
-def build_last_seen_interval_key(detector: Detector) -> str:
-    return f"project-sub-last-seen-interval:detector:{detector.id}"
-
-
-def build_detector_fingerprint_component(detector: Detector) -> str:
-    return f"uptime-detector:{detector.id}"
-
-
-def build_fingerprint(detector: Detector) -> list[str]:
-    return [build_detector_fingerprint_component(detector)]
 
 
 def resolve_uptime_issue(detector: Detector) -> None:
@@ -474,7 +457,7 @@ def disable_uptime_detector(detector: Detector, skip_quotas: bool = False):
             # start from a good state
             detector_state.update(state=DetectorPriorityLevel.OK, is_triggered=False)
 
-        cluster = redis.redis_clusters.get(settings.SENTRY_UPTIME_DETECTOR_CLUSTER)
+        cluster = get_cluster()
         last_update_key = build_last_update_key(detector)
         cluster.delete(last_update_key)
