+++ tests/sentry/uptime/autodetect/test_detector.py
@@ -2,13 +2,14 @@ from sentry.models.organization import Organization
 from sentry.testutils.cases import UptimeTestCase
 from sentry.testutils.helpers.options import override_options
 from sentry.uptime.autodetect.detector import autodetect_base_url_for_project
-from sentry.uptime.autodetect.ranking import _get_cluster, get_organization_bucket_key
+from sentry.uptime.autodetect.ranking import get_organization_bucket_key
+from sentry.uptime.utils import get_cluster
 
 
 class DetectBaseUrlForProjectTest(UptimeTestCase):
     def assert_organization_key(self, organization: Organization, exists: bool) -> None:
         key = get_organization_bucket_key(organization)
-        cluster = _get_cluster()
+        cluster = get_cluster()
         assert exists == cluster.sismember(key, str(organization.id))
 
     def test(self) -> None:
