+++ tests/sentry/uptime/test_grouptype.py
@@ -26,12 +26,9 @@ from sentry.uptime.grouptype import (
     build_evidence_display,
 )
 from sentry.uptime.models import UptimeSubscription, get_uptime_subscription
-from sentry.uptime.subscriptions.subscriptions import (
-    build_detector_fingerprint_component,
-    build_fingerprint,
-    resolve_uptime_issue,
-)
+from sentry.uptime.subscriptions.subscriptions import resolve_uptime_issue
 from sentry.uptime.types import UptimeMonitorMode
+from sentry.uptime.utils import build_detector_fingerprint_component, build_fingerprint
 from sentry.workflow_engine.models.data_source import DataPacket
 from sentry.workflow_engine.models.detector import Detector
 from sentry.workflow_engine.types import DetectorEvaluationResult, DetectorPriorityLevel
